backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
click==8.1.8
dnspython==2.7.0
fastapi==0.109.2
google-generativeai==0.8.4
h11==0.14.0
httpcore==1.0.7
httpx==0.26.0
idna==3.10
iniconfig==2.1.0
jina==3.34.0
motor==3.3.2
packaging==24.2
pluggy==1.5.0
pydantic==2.6.1
pydantic_core==2.16.2
PyPDF2==3.0.1
pymongo==4.6.2
pytest==7.4.4
python-dotenv==1.0.0
python-multipart==0.0.9
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.13.1
uvicorn==0.23.1
requests==2.31.0
grpcio==1.68.0
